Move MockCPU and MockCPUs
