+++ tests/test_monitor.py
@@ -23,9 +23,10 @@ import pynvml
 import pytest
 
 from zeus.monitor import Measurement, ZeusMonitor
-from zeus.utils.testing import ReplayZeusMonitor, MOCKCPUs, NUM_CPUS
+from zeus.utils.testing import ReplayZeusMonitor
 import zeus.device.gpu
 import zeus.device.cpu
+from zeus.device.cpu.common import CpuDramMeasurement, CPUs, CPU
 
 if TYPE_CHECKING:
     from pathlib import Path
@@ -38,6 +39,49 @@ ARCHS = [
     pynvml.NVML_DEVICE_ARCH_VOLTA,
     pynvml.NVML_DEVICE_ARCH_AMPERE,
 ]
+NUM_CPUS = 2
+
+
+class MockCPU(CPU):
+    """Control a single MOCK CPU for testing."""
+
+    def __init__(self, index):
+        """Initialize the MOCKCPU with a specified index for testing."""
+        self.index = index
+        self.cpu_energy = itertools.count(start=1000, step=10)
+        self.dram_energy = (
+            itertools.count(start=200, step=5) if self.index % 2 == 0 else None
+        )
+
+    def getTotalEnergyConsumption(self):
+        """Returns the total energy consumption of the specified powerzone. Units: mJ."""
+        return CpuDramMeasurement(
+            cpu_mj=float(next(self.cpu_energy)),
+            dram_mj=float(next(self.dram_energy))
+            if self.dram_energy is not None
+            else None,
+        )
+
+    def supportsGetDramEnergyConsumption(self):
+        """Returns True if the specified CPU powerzone supports retrieving the subpackage energy consumption."""
+        return self.dram_energy is not None
+
+
+class MockCPUs(CPUs):
+    """MOCK CPU Manager object, containing individual MOCKCPU objects for testing."""
+
+    def __init__(self):
+        """Instantiates MOCKCPUs object for testing."""
+        self._cpus = [MockCPU(i) for i in range(NUM_CPUS)]
+
+    @property
+    def cpus(self) -> Sequence[CPU]:
+        """Returns a list of CPU objects being tracked."""
+        return self._cpus
+
+    def __del__(self) -> None:
+        """Shuts down the Mock CPU monitoring."""
+        return
 
 
 @pytest.fixture(autouse=True, scope="function")
@@ -211,7 +255,7 @@ def test_monitor(pynvml_mock, mock_gpus, mocker: MockerFixture, tmp_path: Path):
 
     # want to make zeus.device.gpu.nvml_is_available is a function, want it to always return true when testing
     mocker.patch("zeus.device.gpu.nvml_is_available", return_value=True)
-    mocker.patch("zeus.device.cpu._cpus", new=MOCKCPUs())
+    mocker.patch("zeus.device.cpu._cpus", new=MockCPUs())
 
     ########################################
     # Test ZeusMonitor initialization.
